BUGFIX: Protect against undefined sightings value
Occasionally Mongo returns an empty set for 'sightings' that match newName's values, despite successfully inserting the newName document immediately preceding. Given Mongo's concurrent nature/lack of built in transactions, I'm thinking this is the result of a race condition.

This if/else logic is a temporary workaround to protect against raising exceptions (calling a function on an undefined variable) while the above suspicion is being researched.